move BSPLanguageSupport into util_rules.lifecycle
[ci skip-rust]

[ci skip-build-wheels]
