+++ src/python/pants/backend/scala/bsp/rules.py
@@ -16,7 +16,7 @@ from pants.bsp.context import BSPContext
 from pants.bsp.protocol import BSPHandlerMapping
 from pants.bsp.rules import BSPBuildTargets, BSPBuildTargetsRequest
 from pants.bsp.spec import BuildTarget, BuildTargetCapabilities, BuildTargetIdentifier
-from pants.bsp.types import BSPLanguageSupport
+from pants.bsp.util_rules.lifecycle import BSPLanguageSupport
 from pants.build_graph.address import AddressInput
 from pants.engine.addresses import Addresses
 from pants.engine.internals.selectors import Get, MultiGet
